add generic mk_sim.py
Generic mk_sim.py is a general mk script that takes a directory as a CLI argument.
It reads the yaml file in that directory, then sets up and runs the sim described in that file.
It saves all files in the given directory.

TODO: I'd like to add a combine_sims flag in the yaml, and the neccessary code to combine them,
with an appropriate noise model.